⚡️ refactor(#10): Standardize exception handling in English
